moved package querying to package infolink/querying
