+++ src/main/java/io/github/infolis/infolink/querying/DaraWebMatcher.java
@@ -1,4 +1,4 @@
-package io.github.infolis.querying;
+package io.github.infolis.infolink.querying;
 
 import io.github.infolis.util.RegexUtils;
 import io.github.infolis.util.URLParamEncoder;
