+++ src/main/java/io/github/infolis/infolink/querying/SearchResultScorer.java
@@ -1,4 +1,4 @@
-package io.github.infolis.querying;
+package io.github.infolis.infolink.querying;
 
 import io.github.infolis.model.TextualReference;
 import io.github.infolis.model.entity.SearchResult;
