+++ src/test/java/io/github/infolis/algorithm/SearchPatternsAndCreateLinksTest.java
@@ -20,8 +20,8 @@ import io.github.infolis.model.Execution;
 import io.github.infolis.model.entity.EntityLink;
 import io.github.infolis.model.entity.InfolisFile;
 import io.github.infolis.model.entity.InfolisPattern;
-import io.github.infolis.querying.DaraHTMLQueryService;
-import io.github.infolis.querying.QueryService;
+import io.github.infolis.infolink.querying.DaraHTMLQueryService;
+import io.github.infolis.infolink.querying.QueryService;
 import io.github.infolis.util.SerializationUtils;
 
 /**
